requirements.txt
Django
gunicorn
whitenoise
psycopg2
python-dotenv
djangorestframework